Removed handling of __DIE__ signal.
I removed handling of the __DIE__ signal because it
was causing problems. Because the signalhandler for
the __DIE__ signal was calling remove_pidfile() the
forks ended up removing the PID-file whenever a
backend timed out. Therefore I added a check to
remove_pidfile() to ensure that only the parent
process is allowed to actually remove the PID-file.

I also addded an END statement to take care of
PID-file cleanup.